separate profiler unit tests from main unit tests
