+++ tests/test_main.py
@@ -61,60 +61,6 @@ def test_is_total_flows_unknown(args, input_type, expected_result):
 
 
 @pytest.mark.parametrize(
-    "cpu_profiler_multiprocess, expected_stop_calls, expected_print_calls",
-    [  # Testcase1: CPU profiler enabled, not multiprocess
-        (False, 1, 1),
-        # Testcase2: CPU profiler enabled, multiprocess
-        (True, 0, 0),
-    ],
-)
-def test_cpu_profiler_release_enabled(
-    cpu_profiler_multiprocess,
-    expected_stop_calls,
-    expected_print_calls,
-):
-    main = ModuleFactory().create_main_obj()
-    main.profilers_manager.cpu_profiler_enabled = True
-    main.profilers_manager.cpu_profiler_multiprocess = (
-        cpu_profiler_multiprocess
-    )
-    main.profilers_manager.cpu_profiler = MagicMock()
-    main.profilers_manager.cpu_profiler_release()
-
-    assert (
-        main.profilers_manager.cpu_profiler.stop.call_count
-        == expected_stop_calls
-    )
-    assert (
-        main.profilers_manager.cpu_profiler.print.call_count
-        == expected_print_calls
-    )
-
-
-def test_cpu_profiler_release_disabled():
-    main = ModuleFactory().create_main_obj()
-    main.profilers_manager.cpu_profiler_enabled = False
-    main.profilers_manager.cpu_profiler_release()
-    assert not hasattr(main.profilers_manager, "memory_profiler")
-
-
-def test_memory_profiler_release_enabled():
-    main = ModuleFactory().create_main_obj()
-    main.profilers_manager.memory_profiler_enabled = True
-    main.profilers_manager.memory_profiler = MagicMock()
-    main.profilers_manager.memory_profiler_release()
-    main.profilers_manager.memory_profiler.stop.assert_called_once()
-
-
-def test_memory_profiler_release_disabled():
-    main = ModuleFactory().create_main_obj()
-    main.profilers_manager.memory_profiler_enabled = False
-    main.profilers_manager.memory_profiler_release()
-
-    assert not hasattr(main.profilers_manager, "memory_profiler")
-
-
-@pytest.mark.parametrize(
     "mode, time_diff, expected_calls",
     [  # Testcase1: Should update stats
         ("interactive", 10, 1),
